Change match_id to bigint

+++ database/models.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS "matches" (
-  "match_id" integer NOT NULL,
+  "match_id" bigint NOT NULL,
   "start_time" timestamp with time zone NOT NULL,
   "lobby_type" integer NOT NULL,
   CONSTRAINT pk_matches PRIMARY KEY ("match_id")
